backend/requirements.txt
Django==4.2.13
psycopg2-binary==2.9.9
djangorestframework==3.15.1
django-cors-headers==4.3.1
